Added context menu for adding new scene objects. Dividied CodeEditor into CodeViewer and CodeEditor.
